Haciendo la carga de saldos

+++ kiosco/views.py
@@ -1,8 +1,11 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.db.models import Q
 from django.db.models.functions import Right 
+from django.core.paginator import Paginator
+from django.contrib import messages
 from kiosco.models import *
 from kiosco.forms import *
+from transacciones.models import *
 from decimal import Decimal
 from django.contrib.auth.decorators import login_required, user_passes_test
 
@@ -18,11 +21,20 @@ def home(request):
 @user_passes_test(lambda u: u.is_superuser)
 def lista_clientes(request):
     busqueda = request.GET.get("nombre") 
+    orden = request.GET.get('orden')
     clientes_query = Cliente.objects.all()
     if busqueda:
         clientes_query = Cliente.objects.filter(
-            Q(nombre__icontains = busqueda) | Q(apellido__icontains=busqueda),
+            Q(nombre__icontains = busqueda) | Q(apellido__icontains=busqueda) | Q(usuario__first_name__icontains=busqueda) | Q(usuario__last_name__icontains=busqueda),
         )
+    
+    if orden == 'apellido':
+        clientes_query = clientes_query.order_by('apellido')
+    elif orden == 'curso':
+        clientes_query = clientes_query.order_by('curso')
+    elif orden == 'tutor':
+        clientes_query = clientes_query.order_by('usuario') # Cambia 'tutor' por el campo real
+
     return render(request, 'kiosco/clientes.html',{"clientes":clientes_query})
 
 @login_required
@@ -42,11 +54,17 @@ def crear_cliente(request):
 @login_required
 def ver_cliente(request, pk):
     cliente = get_object_or_404(Cliente, pk=pk)
+    if cliente.usuario != request.user:
+        messages.error(request, 'No tienes permiso para ver ese cliente.')
+        return redirect('home')
     return render(request, 'kiosco/ver_cliente.html', {"cliente":cliente})
 
 @login_required
 def actualizar_cliente(request, pk):
     cliente = get_object_or_404(Cliente, pk=pk)
+    if cliente.usuario != request.user:
+        messages.error(request, 'No tienes permiso para editar ese cliente.')
+        return redirect('home')
     if request.method == "POST":
         form = ClienteForm(request.POST, instance=cliente)
         if form.is_valid():
@@ -61,8 +79,30 @@ def actualizar_cliente(request, pk):
     })
 
 @login_required
+def historial_cliente(request, pk):
+    cliente = get_object_or_404(Cliente, pk=pk)
+    if cliente.usuario != request.user:
+        messages.error(request, 'No tienes permiso para ver el historial de ese cliente.')
+        return redirect('home')
+    movimientos_list = Transaccion.objects.filter(tarjeta__cliente=cliente).order_by('-fecha')
+    
+    paginator = Paginator(movimientos_list, 20)
+    page_number = request.GET.get('page')
+    movimientos = paginator.get_page(page_number)
+    
+    return render(request, 'kiosco/historial_completo_cliente.html', {
+        'cliente': cliente,
+        'movimientos': movimientos, 
+        'page_obj': movimientos,    
+        'is_paginated': True
+    })
+
+@login_required
 def eliminar_cliente(request, pk):
     cliente = get_object_or_404(Cliente, pk=pk)
+    if cliente.usuario != request.user:
+        messages.error(request, 'No tienes permiso para editar ese cliente.')
+        return redirect('home')
     if request.method == "POST":
         cliente.delete()
         return redirect('home')
@@ -75,6 +115,13 @@ def eliminar_cliente(request, pk):
 def lista_tarjetas(request):
     codigo = request.GET.get("codigo") 
     tarjetas_query = Tarjeta.objects.all()
+
+    filtro = request.GET.get('filtro')
+    if filtro == 'con_usuario':
+        tarjetas_query = tarjetas_query.exclude(cliente__isnull=True)
+    elif filtro == 'sin_usuario':
+        tarjetas_query = tarjetas_query.filter(cliente__isnull=True)
+    
     if codigo:
         tarjetas_query = tarjetas_query.annotate(
             ultimos_digitos=Right('codigo', 9)
@@ -116,6 +163,28 @@ def asociar_tarjeta(request, pk):
         'query': busqueda_cliente
     })
 
+@user_passes_test(lambda u: u.is_superuser)
+def cambiar_estado_tarjeta(request, pk):
+    tarjeta = get_object_or_404(Tarjeta, pk=pk)
+    tarjeta.habilitada = not tarjeta.habilitada 
+    tarjeta.save()
+
+    return redirect('ver_tarjeta', tarjeta.pk)
+
+
+@login_required
+def cambiar_estado_tarjeta_alumno(request, pk):
+    tarjeta = get_object_or_404(Tarjeta, pk=pk)
+    if tarjeta.cliente.usuario != request.user:
+        messages.error(request, 'No puedes modificar esa tarjeta')
+        return redirect('home')
+    tarjeta.habilitada = not tarjeta.habilitada 
+    tarjeta.save()
+
+    return redirect('ver_cliente', pk=tarjeta.cliente.pk)
+
+
+@user_passes_test(lambda u: u.is_superuser)
 def asociar_tarjeta_confirmar(request, pk, cliente_pk):
     tarjeta = get_object_or_404(Tarjeta, pk=pk)
     cliente = get_object_or_404(Cliente, pk=cliente_pk)
@@ -130,30 +199,6 @@ def asociar_tarjeta_confirmar(request, pk, cliente_pk):
         'tarjeta': tarjeta,
         'cliente': cliente
     })
-
-
-
-# @user_passes_test(lambda u: u.is_superuser)
-# def actualizar_tarjeta(request, pk):
-#     tarjeta = get_object_or_404(Tarjeta, pk=pk)
-#     if request.method == "POST":
-#         monto_str = request.POST.get("monto")
-#         if monto_str:
-#             try:
-#                 monto_a_sumar = Decimal(monto_str)
-#                 tarjeta.saldo += monto_a_sumar
-#                 tarjeta.save()
-
-#                 return redirect("ver_tarjeta",pk = tarjeta.pk)
-#             except ValueError:
-#                 form = TarjetaSaldoForm(instance=tarjeta)
-#     else:
-#         form = TarjetaSaldoForm(instance=tarjeta)
-
-#     return render(request, "kiosco/saldo_tarjeta.html",{
-#         "form":form,
-#         "tarjeta":tarjeta        
-#     })
 
 @user_passes_test(lambda u: u.is_superuser)
 def eliminar_tarjeta(request, pk):
